model/requirements.txt
absl-py==0.10.0
astunparse==1.6.3
attrs==20.3.0
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
dill==0.3.3
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-resources==3.3.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==8.0.1
promise==2.3
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow-datasets==4.1.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.3.0
tensorflow-gpu-estimator==2.3.0
tensorflow-metadata==0.26.0
termcolor==1.1.0
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1